RFC: customizable HTML renderer
As of now, djot.js provides extensibility via filters -- code to modify
djot's own AST. This is I think of limited usability: djot is the source
format, not the destination format, so most interesting logic should
happen during the output.

For example, if I want to render `![](url){.video}` as `<video>`, rather
than `<img>`, with filters I'd have to introduce a verbatim node, which
is rather ugly. An even harder case is when I want to replace some tag
while keeping default rendering for children.

I think a nice way to solve that would be to allow customizing HTML
rendering itself, by essentially letting the user to supply a bunch of
overrides for various nodes.

This PR implements a proof of concept, with the call-site looking like
this:

```ts
const ast = djot.parse(`
    # Example

    ![](http://example.com/video.mp4){.video}

    ![](http://example.com/cat.gif)
`);

const html = djot.renderHTML(ast, {
    overrides: {
        image: (node, renderer) => {
            if ((node.attributes?.["class"] ?? "").includes("video")) {
                return renderer.renderTag("video", node, { src: node.destination })
            }
            return renderer.renderAstNodeDefault(node)
        }
    }
})
```