Changed parsing code for links/images.
We now pop the image/link container at -image or -link,
and then when -reference or -destination is encountered,
we modify the last child.

Instead of adding the `accumulatedText` that comes from
str, softbreak, and space elements, we just add the whole
line from the start of the destination or reference.

Closes #46.