test: fix and test number comments
Many comments were misaligned in 9f7ddab (eslint fixes)
by changing `let` to `const` which is two chars longer.

This commit fixes the alignments as well as consistently
puts the digits for `\n` below the blackslash.
A respective test is introduced (which can update all comments
if the OVERRIDE environment variable is set).