`parse.ts` now exports `parseFromEvents` [API change].
This takes an `Event[]` parameter, which can be produced from
an `EventParser` using `Array.from()`.

`parse` is now defined in terms of `parseFromEvents`.

This will help users who need both the parsed content and the
event stream. Previously they had to do the parsing twice, once
using `parseEvents` and once with `parse` (which itself calls
`parseEvents`). This can now be avoided; `parseEvents` can be
called once and then `parseFromEvents` can be called using the
extracted events.

Users of the old `parse` function should not see any changes.

Note: it is possible that there could be performance effects due to the
reification of the Event array, since the previous version of `parse`
consumed Events as they were streamedfrom `parseEvents` rather than
taking an array of all the Events as parameter. However, in our own
benchmarks we don't see any significant differences.

Closes #111.